WIP Start making alternative IRs viewable/interactable at the function level
The purely block-level view is nice, but isn't complete because there isn't
always a 1-1 mapping between blocks in different IRs.  It will be useful to be
able to look at functions in different IRs independent of the block mapping.

This change will modify the function viewer to be able to select the current IR
much like the improved block viewer.  The next steps in this change will be to
replace the old Context type with a new GADT (defined already).  The current
context will either be a function or block.  The context can be swapped between
them trivially.  This will remove a lot of partiality, because there won't be
maps keyed on IRRepr anymore - the current repr/view will be reflected in the
context, and switching will just change the context.

This will also let us reflect changes in function view in the context stack.